try_lower_bind_unwrap_return takes its admission verdict and both arms from named helpers, holding the certificate spine
