+++ crates/almide-mir/src/lower/binds_b.rs
@@ -589,6 +589,7 @@ impl LowerCtx {
 }
 
 include!("binds_p2.rs");
+include!("binds_p2_bang.rs");
 include!("binds_p2_heap.rs");
 include!("binds_p2_b.rs");
 include!("binds_p2_c.rs");
